requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.1
Django==4.2.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
gunicorn==21.2.0
iniconfig==2.0.0
kombu==5.3.2
packaging==23.1
Pillow==9.5.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
pytest==7.4.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
